==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_raw_data(year:int,month:int)->Path:
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(URL)

    if response.status_code == 200:
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        open(path,'wb').write(response.content)
        return path
    else:
        raise Exception(f"{URL} is not available")

download_raw_data(2024,1)

import pandas as pd

rides = pd.read_parquet("../data/raw/rides_2024-01.parquet")
rides.head(20)


rides = rides[["tpep_pickup_datetime","PULocationID"]]

## Rename column name

rides.rename(columns={
    "tpep_pickup_datetime":"pickup_datetime",
    "PULocationID":"pickup_location_id"
},inplace=True)

rides.head(20)


rides['pickup_datetime'].describe()

# find min and max pickup datetime
rides["pickup_datetime"].max()

rides["pickup_datetime"].max()

## Want only data of JANUARY
## Data Validation  

rides = rides[rides["pickup_datetime"] >= '2024-01-01']
rides = rides[rides["pickup_datetime"] < '2024-02-01']
rides['pickup_datetime'].describe()

## Store validated data

rides.to_parquet("../data/transformed/validated_rides_2024_01.parquet")
