==> notebooks/04_transforme_raw_data_into_features_and_targets.ipynb <==
%reload_ext autoreload
%autoreload 2

from data import load_raw_data

rides = load_raw_data(2024)
rides.head()

from data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data.head()

from data import transform_raw_data_into_features_and_targets

features, targets = transform_raw_data_into_features_and_targets(ts_data)
features.head()
targets.head()

from data import transform_raw_data_into_features_and_targets

features, targets = transform_raw_data_into_features_and_targets(
    ts_data=ts_data,
    input_seq_length=24 * 28 * 1, # 1 month
    step_size=24, # 1 day
)

print(f"Features shape: {features.shape}")
print(f"Targets shape: {targets.shape}")


tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
