==> notebooks/06_baseline_model.ipynb <==
%reload_ext autoreload
%autoreload 2

# load training data
import pandas as pd
from paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

## Split into training and test

from datetime import datetime
from data_split import train_test_split

cutoff_date = datetime(2024, 6, 1,0,0,0)
train_X, train_y, test_X, test_y = train_test_split(df, cutoff_date, target_column_name='target_rides_next_hour')

print(f'train_X shape: {train_X.shape}')
print(f'train_y shape: {train_y.shape}')
print(f'test_X shape: {test_X.shape}')
print(f'test_y shape: {test_y.shape}')

test_X.head()


import numpy as np

class BaselineModelPreviousHour:
    """
    Prediction = actual demand observed in the last hour
    """
    def __init__(self):
        pass

    def fit(self, train_X: pd.DataFrame, train_y: pd.Series):
        pass
    
    def predict(self, test_X) -> np.array:
        return test_X['rides_previous_1_hour']


model = BaselineModelPreviousHour()
predictions = model.predict(test_X)
predictions

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(test_y,predictions)
print(f'{test_mae}=:.4f')

import numpy as np
import pandas as pd

class BaselineModelPreviousWeek:
    """
    Prediction = actual demand observed in the last week
    """
    def __init__(self):
        pass
    
    def fit(self, train_X: pd.DataFrame, train_y: pd.Series):
        pass
    
    def predict(self, test_X) -> np.array:
        return test_X[f'rides_previous_{7*24}_hour']


model = BaselineModelPreviousWeek()
predictions = model.predict(test_X)

## Finding Smarter Baseline

test_mae = mean_absolute_error(test_y,predictions)
print(f'{test_mae}=:.4f')

class BaselineModelLast4Weeks:
    """
    Prediction = actual demand observed in the last 4 weeks
    """
    def __init__(self):
        pass
    
    def fit(self, train_X: pd.DataFrame, train_y: pd.Series):
        pass
    
    def predict(self, test_X) -> np.array:
        return 0.25 * (
            test_X[f'rides_previous_{7*24}_hour'] + 
            test_X[f'rides_previous_{2*7*24}_hour'] + 
            test_X[f'rides_previous_{3*7*24}_hour'] + 
            test_X[f'rides_previous_{4*7*24}_hour']
        )


model = BaselineModelLast4Weeks()
predictions = model.predict(test_X)

test_mae = mean_absolute_error(test_y,predictions)
print(f'{test_mae}=:.4f')
