==> notebooks/07_xgboost_model.ipynb <==
import pandas as pd
from paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df.head()

from datetime import datetime

from data_split import train_test_split

train_X, train_y, test_X, test_y = train_test_split(
    df,
    cutoff_date=datetime(2024, 6, 1,0,0,0),
    target_column_name="target_rides_next_hour",
)

print(f'train_X.shape: {train_X.shape}')
print(f'train_y.shape: {train_y.shape}')
print(f'test_X.shape: {test_X.shape}')
print(f'test_y.shape: {test_y.shape}')

import xgboost as xgb

# use only past rides data
past_rides_columns = [c for c in train_X.columns if c.startswith('rides_')]
train_X_only_numeric = train_X[past_rides_columns]

model = xgb.XGBRegressor()
model.fit(train_X_only_numeric, train_y)


test_X_only_numeric = test_X[past_rides_columns]
predictions = model.predict(test_X_only_numeric)
